refactor IO routines into lib
